Remove Types namespace
