+++ model/Config/InternalOrganisation.ts
@@ -2,12 +2,13 @@ import { isly } from "isly"
 import { OrganisationFlag } from "../OrganisationFlag"
 import { PaxpayFeature } from "../PaxpayFeature"
 import { TierID } from "../TierID"
-import { Types } from "./Types"
+import { FundingLimit } from "./FundingLimit"
 
+// InternalOrganisationConfig
 export interface InternalOrganisation {
 	flags?: OrganisationFlag[]
 	features?: PaxpayFeature[]
-	fundingLimitConfig?: Types.FundingLimit
+	fundingLimitConfig?: FundingLimit
 	tier?: TierID
 }
 
@@ -15,7 +16,7 @@ export namespace InternalOrganisation {
 	export const type = isly.object<InternalOrganisation>({
 		flags: OrganisationFlag.type.array().optional(),
 		features: PaxpayFeature.type.array().optional(),
-		fundingLimitConfig: Types.FundingLimit.type.optional(),
+		fundingLimitConfig: FundingLimit.type.optional(),
 		tier: TierID.type.optional(),
 	})
 	export const is = type.is
